refactor(software-demo): move SHA2 plumbing into own module
