+++ embedded-cal-software-demo/src/lib.rs
@@ -9,6 +9,7 @@
 use embedded_cal::{Cal, accessor::*, plumbing::Plumbing};
 
 mod hash;
+mod hash_plumbing;
 mod hkdf;
 mod hmac;
 mod rng;
@@ -69,29 +70,6 @@ impl<EC: ExtenderConfig> embedded_cal::plumbing::ec::Ec for Extender<EC> {
 
     fn x448(&mut self) -> &mut Self::PrimitivesX448 {
         self.0.x448()
-    }
-}
-
-impl<EC: ExtenderConfig> embedded_cal::plumbing::hash::Sha2Short for Extender<EC> {
-    const SUPPORTED: bool = <EC::Base as embedded_cal::plumbing::hash::Sha2Short>::SUPPORTED;
-    const SEND_PADDING: bool = <EC::Base as embedded_cal::plumbing::hash::Sha2Short>::SEND_PADDING;
-    const FIRST_CHUNK_SIZE: usize =
-        <EC::Base as embedded_cal::plumbing::hash::Sha2Short>::FIRST_CHUNK_SIZE;
-    const UPDATE_MULTICHUNK: bool =
-        <EC::Base as embedded_cal::plumbing::hash::Sha2Short>::UPDATE_MULTICHUNK;
-
-    type State = <EC::Base as embedded_cal::plumbing::hash::Sha2Short>::State;
-
-    fn init(&mut self, variant: embedded_cal::plumbing::hash::Sha2ShortVariant) -> Self::State {
-        self.0.init(variant)
-    }
-
-    fn update(&mut self, instance: &mut Self::State, data: &[u8]) {
-        self.0.update(instance, data)
-    }
-
-    fn finalize(&mut self, instance: Self::State, last_chunk: &[u8], target: &mut [u8]) {
-        self.0.finalize(instance, last_chunk, target)
     }
 }
